Latest Update

+++ Libft/Stdlib/ft_ftoa.c
@@ -6,14 +6,14 @@
 /*   By: akharrou <akharrou@student.42.us.org>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/17 16:20:16 by akharrou          #+#    #+#             */
-/*   Updated: 2019/04/18 13:15:04 by akharrou         ###   ########.fr       */
+/*   Updated: 2019/04/30 10:26:18 by akharrou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../Includes/stdlib_42.h"
-#include "../Includes/macros.h"
+#include "../Includes/macros_42.h"
 
-char	*ft_ftoa(long double n, int width, int precision)
+char	*ft_ftoa(float n, int width, int precision)
 {
 	return (ft_ftoa_base(n, DECIMAL_BASE, width, precision));
 }
